# public/swagger.json
{
  "components": {
    "examples": {},
    "headers": {},
    "parameters": {},
    "requestBodies": {},
    "responses": {},
    "schemas": {
      "FlattenMaps_T_": {
        "properties": {},
        "type": "object"
      },
      "IList": {
        "$ref": "#/components/schemas/FlattenMaps_T_"
      },
      "IUser": {
        "$ref": "#/components/schemas/FlattenMaps_T_"
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "info": {
    "title": "shoppieapi",
    "version": "1.0.0",
    "license": {
      "name": "ISC"
    },
    "contact": {}
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "openapi": "3.0.0",
  "paths": {
    "/lists": {
      "get": {
        "operationId": "FindByUserId",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/IList"
                  },
                  "type": "array"
                }
              }
            }
          }
        },
        "security": [],
        "parameters": []
      }
    },
    "/lists/{id}": {
      "get": {
        "operationId": "FindById",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/IList"
                    }
                  ],
                  "nullable": true
                }
              }
            }
          }
        },
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/user/register": {
      "post": {
        "operationId": "Register",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/IUser"
                    },
                    {
                      "properties": {
                        "error": {
                          "type": "string"
                        }
                      },
                      "required": ["error"],
                      "type": "object"
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "password": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  }
                },
                "required": ["password", "email"],
                "type": "object"
              }
            }
          }
        }
      }
    },
    "/user/login": {
      "post": {
        "operationId": "Login",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/IUser"
                    },
                    {
                      "properties": {
                        "error": {
                          "type": "string"
                        }
                      },
                      "required": ["error"],
                      "type": "object"
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "password": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  }
                },
                "required": ["password", "email"],
                "type": "object"
              }
            }
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}